crud category && update course

+++ src/main/java/com/example/backendkhoaluan/service/CourseServiceImp.java
@@ -1,9 +1,11 @@
 package com.example.backendkhoaluan.service;
 
+import com.example.backendkhoaluan.constant.Constants;
 import com.example.backendkhoaluan.dto.CoursesDTO;
 import com.example.backendkhoaluan.entities.Courses;
 import com.example.backendkhoaluan.entities.RatingCourse;
 import com.example.backendkhoaluan.entities.User;
+import com.example.backendkhoaluan.exception.DataNotFoundException;
 import com.example.backendkhoaluan.exception.DeleteException;
 import com.example.backendkhoaluan.exception.InsertException;
 import com.example.backendkhoaluan.payload.request.CreateCourseRequest;
@@ -33,39 +35,43 @@ public class CourseServiceImp implements CourseService {
     private FilesStorageService filesStorageService;
 
     @Override
-    public Page<Courses> getAllCourse(CustomCourseQuery.CourseFilterParam param,PageRequest pageRequest) {
+    public Page<Courses> getAllCourse(CustomCourseQuery.CourseFilterParam param, PageRequest pageRequest) {
         Specification<Courses> specification = CustomCourseQuery.getFilterCourse(param);
         return coursesRepository.findAll(specification, pageRequest);
     }
 
     @Override
-    @Transactional(rollbackFor = {InsertException.class, Exception.class})
+    @Transactional(rollbackFor = {DeleteException.class, Exception.class})
     public void deleteCourse(int id) {
         try {
-            Optional<Courses> coursesOptional=coursesRepository.findById(id);
-            coursesOptional.ifPresent(data->{
-                filesStorageService.deleteAll(data.getImage());
-                coursesRepository.delete(data);
-            });
-        }catch (Exception e){
-            throw new DeleteException("Xóa khóa học thất bại",e.getLocalizedMessage());
+            Optional<Courses> coursesOptional = coursesRepository.findById(id);
+            if (!coursesOptional.isPresent()) {
+                throw new DataNotFoundException(Constants.ErrorMessageCourseValidation.NOT_FIND_COURSE_BY_ID + id);
+            }
+            Courses data = coursesOptional.get();
+            filesStorageService.deleteAll(data.getImage());
+            coursesRepository.delete(data);
+        } catch (Exception e) {
+            throw new DeleteException("Xóa khóa học thất bại", e.getLocalizedMessage());
         }
     }
 
     @Override
     public CoursesDTO getCourseById(int id) {
-        Optional<Courses> coursesOptional=coursesRepository.findById(id);
+        Optional<Courses> coursesOptional = coursesRepository.findById(id);
+        if (!coursesOptional.isPresent()) {
+            throw new DataNotFoundException(Constants.ErrorMessageCourseValidation.NOT_FIND_COURSE_BY_ID + id);
+        }
+        Courses data = coursesOptional.get();
         CoursesDTO courseDTO = new CoursesDTO();
-        coursesOptional.ifPresent(data->{
-            courseDTO.setId(data.getId());
-            courseDTO.setDescription(data.getDescription());
-            courseDTO.setName(data.getName());
-            courseDTO.setPrice(data.getPrice());
-            courseDTO.setDiscount(data.getDiscount());
-            courseDTO.setImage("http://localhost:8081/api/file/"+data.getImage());
-            courseDTO.setIdUser(data.getUser().getId());
-            courseDTO.setRating(calculatorRating(data.getListRatingCourses()));
-        });
+        courseDTO.setId(data.getId());
+        courseDTO.setDescription(data.getDescription());
+        courseDTO.setName(data.getName());
+        courseDTO.setPrice(data.getPrice());
+        courseDTO.setDiscount(data.getDiscount());
+        courseDTO.setImage("http://localhost:8081/api/file/" + data.getImage());
+        courseDTO.setIdUser(data.getUser().getId());
+        courseDTO.setRating(calculatorRating(data.getListRatingCourses()));
 
         return courseDTO;
     }
@@ -74,9 +80,36 @@ public class CourseServiceImp implements CourseService {
     @Override
     public void save(CreateCourseRequest createCourseRequest, MultipartFile file) {
         try {
-            String fileName=filesStorageService.save(file);
+            String fileName = filesStorageService.save(file);
 
             Courses courseEntity = new Courses();
+            courseEntity.setName(createCourseRequest.getName());
+            courseEntity.setDescription(createCourseRequest.getDescription());
+            courseEntity.setDiscount(createCourseRequest.getDiscount());
+            courseEntity.setImage(fileName);
+            courseEntity.setPrice(createCourseRequest.getPrice());
+
+            User userEntity = new User();
+            userEntity.setId(createCourseRequest.getUserId());
+
+            courseEntity.setUser(userEntity);
+
+            coursesRepository.save(courseEntity);
+        } catch (Exception e) {
+            throw new InsertException("Lỗi insert course", e.getLocalizedMessage());
+        }
+    }
+
+    @Transactional(rollbackFor = {InsertException.class, Exception.class})
+    @Override
+    public void updateCourse(int id,CreateCourseRequest createCourseRequest, MultipartFile file) {
+        try {
+            Optional<Courses> coursesOptional = coursesRepository.findById(id);
+            if (!coursesOptional.isPresent()) {
+                throw new DataNotFoundException(Constants.ErrorMessageCourseValidation.NOT_FIND_COURSE_BY_ID + id);
+            }
+            Courses courseEntity=coursesOptional.get();
+            String fileName = filesStorageService.save(file);
             courseEntity.setName(createCourseRequest.getName());
             courseEntity.setDescription(createCourseRequest.getDescription());
             courseEntity.setDiscount(createCourseRequest.getDiscount());
